requirements.txt
sentence-transformers
yake
spacy
pytextrank
python-dotenv
chromadb
google-generativeai
Flask
Flask-Cors
watchdog